Add files via upload
1. Write code for the following problem: calculates the biggest value of an unsorted array with unique elements. You can get the number of elements the user enters before the user enter the elements of the array. Then you must get the array as an input.
2. Write code to calculate the second smallest number in an array of n elements where the elements of the array and n are given by the user.
3. Create a string, called str, of size 20. Then ask the user to enter a word. Scan the user input and save it in the string. This can be done by following scanf command: scanf(“%s”, str);
Print the world user entered. Now calculate the number of alphabets in the world that the user entered. For example, if the user enters “Hello”, the number of alphabets in the word is 5. You must print this as the output. You are not allowed to use any other libraries rather than stdio.h and math.h!